+++ apps/learn-card-app/src/pages/skills/SkillsPage.tsx
@@ -62,7 +62,6 @@ const SkillsPage: React.FC = () => {
     const { backgroundSecondaryColor } = colors;
 
     const flags = useFlags();
-    const showAiInsights = flags?.showAiInsights;
     const showAdminPanel =
         flags?.enableAdminTools ||
         lcNetworkProfile?.role === LearnCardRolesEnum.admin ||
@@ -127,7 +126,6 @@ const SkillsPage: React.FC = () => {
                 <IonContent fullscreen className="skills-page" color={backgroundSecondaryColor}>
                     <div className="flex relative justify-center items-center w-full pb-[30px]">
                         <div className="w-full max-w-[600px] flex items-center justify-center flex-wrap text-center ion-padding mt-[30px] px-[20px]">
-                            {/* {showAiInsights && <SkillsInsightCard />} */}
                             {showAdminPanel && (
                                 <div
                                     className={`flex items-center justify-start w-full ${