+++ apps/scouts/src/helpers/externalLinkHelpers.ts
@@ -4,7 +4,6 @@ import { Capacitor } from '@capacitor/core';
 export const TOS_LINK = 'https://www.scout.org/additional-terms-user-accounts';
 export const PP_LINK = 'https://www.scout.org/privacy';
 export const CONTACT_LINK = 'https://support.scout.org/hc/en-gb';
-export const MV_TYPEFORM = 'https://r18y4ggjlxv.typeform.com/to/fBWigxQx'; // Request Metaversity at Your School
 export const LEARNCARD_WEBSITE = 'https://scout.org/scoutpass';
 
 export const openExternalLink = async (url: string) => {