+++ packages/plugins/claimable-boosts/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @learncard/claimable-boosts-plugin
 
+## 1.1.30
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.4.30
+
 ## 1.1.29
 
 ### Patch Changes